+++ src/components/views/Dashboard.js
@@ -6,27 +6,8 @@ import EventItemSquare from "../ui/EventItemSquare";
 import "../../styles/_theme.scss";
 import Header from "./Header";
 import "../../styles/views/Dashboard.scss";
-import {MyButton} from "../ui/MyButton";
-import {apiLoggedIn, handleError} from "../../helpers/api";
-import {useNavigate} from "react-router";
 
 const Dashboard = (props) => {
-
-  const navigate = useNavigate();
-  let userId = localStorage.getItem('userId');
-
-  const logout = async () => {
-    try {
-      await apiLoggedIn().put(`/logout/${userId}`);
-      // Remove the token from the local storage.
-      localStorage.removeItem('token');
-      localStorage.removeItem('userId');
-      navigate('/');
-
-    } catch (error) {
-      alert(`Something went wrong during logout: \n${handleError(error)}`);
-    }
-  };
 
   return (
     <>
@@ -50,7 +31,6 @@ const Dashboard = (props) => {
           </div>
         </div>
       </div>
-      <MyButton onClick={() => logout()}>Logout</MyButton>
       </>
   );
 };
