+++ src/styles/ui/MyButton.scss
@@ -1,5 +1,4 @@
 .MyButton {
-  border: black 1px;
   background-color: #29637c;
   color: white;
   width: 80%;
