+++ src/styles/views/Header.scss
@@ -2,3 +2,17 @@
   background-color: rgb(226, 240, 246);
   height: 100px;
 }
+
+.navbar-brand {
+  margin-left: 3em;
+}
+
+.MyButton {
+  background-color: #29637c;
+  color: white;
+  width: 10% !important;
+  margin-right: 5em;
+  height: 30px;
+  border-radius: 3px;
+  font-size: 10pt;
+}
